feat: collection list 구현
